Point changes

+++ src/TeamTactics.Application/Points/PointResultDto.cs
@@ -6,4 +6,4 @@ using System.Threading.Tasks;
 
 namespace TeamTactics.Application.Points;
 
-public sealed record PointResultDto(string playerName, string clubName, string pointCategoryName, int occurrences, int totalPoints);
+public sealed record PointResultDto(string playerName, string clubName, string pointCategoryName, int occurrences, decimal totalPoints);
